MPC Walkgen: implementation of solve function and corrections Redmine #13154
Reviewed-on: http://gerrit.aldebaran.lan/25469
Tested-by: gerrit
Tested-by: mdegourcuff <mdegourcuff@aldebaran-robotics.com>

+++ test/test-zebulon-com-constraint.cpp
@@ -122,7 +122,7 @@ TEST_F(ZebulonComConstraintTest, sizeOfvalues)
 
   ComConstraint ctr(m1, m2);
 
-  size_t M = m2.getComSupportConvexPolygon().getNbVertices();
+  int M = m2.getComSupportConvexPolygon().getNbVertices();
 
   VectorX jerkInit(4*nbSamples);
   jerkInit.fill(0.0);
